// admin/src/api/generated/schemas/adminVaultAddressBalanceItem.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */

export type AdminVaultAddressBalanceItem = {
  currency?: string;
  contract_address?: string;
  balance?: string;
  balance_usd?: string;
  last_synced_at?: string;
};