// admin/src/api/generated/schemas/adminVaultAddressItem.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */

export type AdminVaultAddressItem = {
  id?: string;
  network_id?: string;
  network?: string;
  address?: string;
  address_type?: string;
  label?: string;
  status?: string;
  is_active_wallet?: boolean;
  total_balance_usd?: string;
  currencies_count?: number;
  last_synced_at?: string;
  created_at?: string;
  created_by_name?: string;
};