// admin/src/api/generated/schemas/adminVaultNetworkOverviewItem.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */
import type { AdminVaultBalanceOverviewItem } from './adminVaultBalanceOverviewItem';
import type { AdminVaultAlert } from './adminVaultAlert';

export type AdminVaultNetworkOverviewItem = {
  network?: string;
  chain_id?: string;
  vault_address?: string;
  status?: string;
  total_balance_usd?: string;
  balances?: AdminVaultBalanceOverviewItem[];
  last_sync_at?: string;
  sync_status?: string;
  alert?: AdminVaultAlert;
  network_id?: string;
};